+++ src/me/zombie_striker/qg/handlers/reloaders/PumpactionReloader.java
@@ -26,7 +26,7 @@ public class PumpactionReloader implements ReloadingHandler {
 
 	@Override
 	public double reload(final Player player, Gun g, int amountReloading) {
-		double time = ((double) g.getReloadTime()) / g.getMaxBullets();
+		double time = (g.getReloadTime()) / g.getMaxBullets();
 		double time2 = time * amountReloading;
 		for (int i = 0; i < amountReloading; i++) {
 			final boolean k = (i + 1 == amountReloading);
